Fix stack buffer overflows in exif_content_dump and exif_entry_dump.
If too large an indent is given, a local buffer will overflow. This
can't happen when called through exif_data_dump (which is likely the
most common case) and since they are documented as being for diagnostic
purposes only, this shouldn't pose too big a security risk in the wild.

Reported-by: jonnygrant.

(not exploitable by malicious data)